fix(ons-page): fix status bar fill for iOS and ons-bottom-toolbar
The fix added for [iOS9](https://github.com/OnsenUI/OnsenUI/commit/eec03eeb6624f995ada9e949659c06f1b553b3d1)
causes an issue if an `ons-bottom-toolbar` is being used (either with an
`ons-toolbar` or stand alone).

When an `ons-bottom-toolbar` is being used the CSS selectors for
adjusting the page header to account for the iOS status bar are no
longer applied because an extra `<div />` is added to the DOM.

Fixes #1047.