fix(ons-page): Use querySelector instead of findChildRecursively.
findChildRecursively has also been removed. It was causing a bug in a very specific scenario. When a page is pushed to the navigator it tries to update it's backbutton which it found with this method. However if the page has an svg element (generated by ons-progress-circular for example) on iOS 8.4 safari - it turns out that the svg doesn't have a children property, causing an error which stops further execution thus omitting the show event. The error is nowhere to be found because of a catch in pushPage.