fix(ons-pull-hook): Fix e2e tests.
Unfortunately the e2e tests don't seem to be working well with the option `dragLockToAxis` as their drag events do not provide initial momentum, causing the dragstart event to always have `direction == right`. This doesn't seem to be affecting the carousel tests since we're using horizontal ones, but for the pullHook it's a problem. Since we want to use this setting and I wasn't able to find a way to make the dragging start with proper momentum this patch just disables the option during the tests instead