fix(ons-navigator, ons-tabbar, ons-splitter): Fix #1209. Start using `$templateCache`.
With this the angular 1 bindings change the templateCache of the core with $templateCache. With this in theory we should be able to get rid of all the custom logic which we have for the templates in the bindings. However just to be safe from some timing issue I have left most of the logic in place.