fix(ons-page): Page now sets its own class if it has a bottom toolbar
Fixes #2459. In this issue, the Vue bindings were causing "page-with-bottom-toolbar" to
be removed. This was because this class was being set on ons-page by ons-bottom-toolbar,
and not being restored when Vue overwrote the classes. If possible, the page should be
responsible for its own classes, so I have moved this class handling from ons-bottom-toolbar
into ons-page.