fix(ons): Fix setImmediate not defined bug for ES modules
ons.internal was calling a polyfill before it was created. This happens
because ons is an external import so it is hoisted, but setup is inlined
at build time, so the ons code was running before the setup code.

This was supposed to be fixed by #2735 but that fix only worked for UMD,
not for ESM.