chore(vue): Fix CSS not being loaded in examples project
This works by importing the CSS from the onsenui package instead of from
a relative local path. At the moment, this means that local CSS changes
won't be reflected in the examples project. To work around this, link
the build directory using `npm link` - this will symlink the onsenui
package to the local build in npm. If the Onsen UI core is restructured
to allow it to be used as an npm workspace, we can drop the need to use
`npm link`.