fix(vue): Stop trying to access props argument in props default function
eslint was throwing an error about calling `this` in VOnsNavigator's
`popPage` prop's default function. I mistakenly thought I fixed that in
587951d8ac by changing `this` to `props`. However, the default function
_does_ actually have access to `this` and does _not_ have `props` passed
in as an argument. The confusion was that default _factory_ functions do
not have access to `this` but instead have `props` passed in as an
argument; default factory functions are for props of type Object. But
`popPage` is of type Function, and so has a default function which is
not a factory function.

I believe this also meanings that eslint is throwing a false error
because it is not checking if `type: Function` is set on the prop. If
`type: Function` is set on the prop, its default function has access to
`this`. If it is not set, then I think Vue assumes it is `type: Object`,
so the default is a factory function, which does not have access to
`this`.