fix(vue): Stop navigator listeners firing twice for ons-page events
This bug was caused by the change in behaviour in Vue 3 where listeners
that are not listed in a component's `emits` key are automatically
applied to the root element. This was a problem when there was a
listener on v-ons-navigator for an event fired by the native ons-page
e.g. 'show'. The listener callback was being called twice when the event
fired: once for the listener that was automatically applied to the
native ons-navigator; and once for the listener that was being applied
to the native ons-page.

This fix sets `inheritAttrs: false` on the pages in the page stack to
stop listeners being automatically applied to the native ons-page.