chore(*): Use Vue 3 instead of Vue 2 in the monorepo
It is currently not possible to have both the Vue 2 and Vue 3 bindings
in the monorepo since they both have the same package name. This commit
can be reverted when we have a solution to that problem.

For now, the Vue 2 bindings can be built by using npm link to link
local versions of onsenui to vue-onsenui, and onsenui and vue-onsenui to
vue-onsenui-examples.